y vectors should be integers

+++ src/main/scala/smltk/naivebayes/NaiveBayes.scala
@@ -6,11 +6,11 @@ trait NaiveBayes {
   private val ZERO_PROBABILITY = 1e-12
   var probabilities = None
 
-  def fit(X: DenseMatrix[Double], y: DenseVector[Double]): Unit
+  def fit(X: DenseMatrix[Double], y: DenseVector[Int]): Unit
 
-  def predict(X: DenseMatrix[Double]): DenseVector[Double]
+  def predict(X: DenseMatrix[Double]): DenseVector[Int]
 
-  def score(Xtest: DenseMatrix[Double], yTrue: DenseVector[Double]): Double = {
+  def score(Xtest: DenseMatrix[Double], yTrue: DenseVector[Int]): Double = {
     val yPreds = predict(Xtest)
     import smltk.metrics.ClassificationMetrics.accuracy
     accuracy(yTrue, yPreds)
